fix: post-rebase hook rename and synthetic click re-entrancy
- Rename useEscapeKey/useOpenState to asEscapeKey/asOpenState to match main
- Fix web-component adapter click routing: only MouseEvent triggers press.commit,
  preventing CustomEvent('click') from asButton's event.emit() from causing
  double-toggle in dropdown and other trigger-based components
