+++ pnpm-workspace.yaml
@@ -3,3 +3,6 @@ packages:
   - 'packages/*/*'
   - 'apps/*'
   - 'internal'
+allowBuilds:
+  esbuild: true
+  sharp: true
